feat: Eden voxel integration — VCET procedural noise port, ISPC SIMD, planet graph nodes
- VoxelProceduralNoise: 17-type stylized noise collection ported from the VCET
  Unreal plugin (MIT, from @lumiey's Procedural Noise Collection), single-source
  impl header compiling as both scalar C++ and ISPC SIMD
- ProceduralNoise2D/3D generator graph nodes (CPU pipeline, range analysis)
- ISPC build integration (ispc_builder.py) with scalar fallback; search path
  updated for C:\DEV_DRIVE\ispc after drive reorg
- Planet/materials graph nodes and engine integration changes for GodotEden

Verified: headless smoke test — all 17 types pass finite/range, determinism,
seed decorrelation, ISPC-vs-scalar parity (max err 0.57% amplitude on Wavelet).

Story: TR-pgen-001 (planet generation pipeline)

+++ meshers/transvoxel/transvoxel_materials_common.h
@@ -6,7 +6,12 @@
 namespace zylann::voxel::transvoxel::materials {
 
 inline uint32_t pack_bytes(const FixedArray<uint8_t, 4> &a) {
-	return (a[0] | (a[1] << 8) | (a[2] << 16) | (a[3] << 24));
+        // Little-endian: a[0] in the least-significant byte.
+        // Matches decode4u8 in all voxel shaders: uvec4(i&0xff, i>>8&0xff, i>>16&0xff, i>>24&0xff)
+        return (static_cast<uint32_t>(a[0]) |
+                        (static_cast<uint32_t>(a[1]) << 8) |
+                        (static_cast<uint32_t>(a[2]) << 16) |
+                        (static_cast<uint32_t>(a[3]) << 24));
 }
 
 inline void add_4i8_4w8_texture_data(
